Backend/requirements.txt
# General modules
Django==4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow


# Third party modules
markdown      
django-filter  
drf-yasg[validation]
django-robots

# deploy
# email third party
django-mail-templated